+++ Doctrine/Phpcr/MultilangMenuNode.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Symfony\Cmf\Bundle\MenuBundle\Doctrine\Phpcr;
-
-use Symfony\Cmf\Bundle\MenuBundle\Model\MultilangMenuNode as BaseMultilangMenuNode;
-
-class MultilangMenuNode extends BaseMultilangMenuNode
-{
-}